requirements.txt
autogen-agentchat
autogen-ext[openai]
python-dotenv
pyowm
aiohttp